reset sketch
move flock simulation to own file
